Modify project structure and add .flaskenv
