+++ .flaskenv
@@ -1,0 +1,2 @@
+FLASK_APP=satchecker.py
+FLASK_DEBUG=0
